+++ src/main/java/com/studentcloud/dbaccess/repo/UniversityRepo.java
@@ -3,7 +3,7 @@ package com.studentcloud.dbaccess.repo;
 import com.studentcloud.dbaccess.entities.University;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UniversityRepo extends CrudRepository<University, Integer> {
+public interface UniversityRepo extends CrudRepository<University, Long> {
 
     boolean existsByName(String name);
